Adding stubbed out api browser plugins.

+++ demo/app/plugins/help/deactivate.js
@@ -4,7 +4,7 @@ var grasshopperInstance = require('../../grasshopper/instance'),
     getTabsContentTypeId = require('../settings').getTabsContentTypeId;
 
 module.exports = function deactivate() {
-    console.log('called deactivate on the Help plugin');
+    console.log(`Called deactivate on the ${require('./config').title} plugin`);
 
     return _queryForThisPluginsTab()
             .then(_deleteThisPluginsTab);
@@ -33,12 +33,16 @@ function _queryForThisPluginsTab() {
 function _deleteThisPluginsTab(queryResults) {
     var found = queryResults.results.find(function(result) { return result.fields.title === require('./config').title; });
 
-    return grasshopperInstance
-            .request
-            .content
-            .deleteById(found._id)
-            .then(function() {
-                console.log('Finished calling deactivate on this plugin.');
-                return { 'state' : 'unactivated' };
-            });
+    if(found) {
+        return grasshopperInstance
+                .request
+                .content
+                .deleteById(found._id)
+                .then(function() {
+                    console.log('Finished calling deactivate on this plugin.');
+                    return { 'state' : 'unactivated' };
+                });
+    } else {
+        return true;
+    }
 }
